% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRV.R
\name{getLRV}
\alias{getLRV}
\title{The getLRV function}
\usage{
getLRV(
 
    mySketch = "http://data.waterpathogens.org/dataset/a1423a05-7680-4d1c-8d67-082fbeb00a50/resource/e7852e8f-9603-4b19-a5fa-9cb3cdc63bb8/download/sketch_lubigi.json",
 
    myLRVdata = "http://data.waterpathogens.org/dataset/eda3c64c-479e-4177-869c-93b3dc247a10/resource/9e172f8f-d8b5-4657-92a4-38da60786327/download/treatmentdata.csv",
  pathogenType = "Virus",
  inFecalSludge = 1e+10,
  inSewage = 1e+10
)
}
\arguments{
\item{pathogenType}{Pathogen group of interest (Virus, Bacteria, Protozoa, Helminth)}

\item{inFecalSludge}{Number of pathogens conveyed each year to the treatment plant in fecal sludge}

\item{inSewage}{Number of pathogens conveyed each year to the treatment plant in sewerage}

\item{sketch}{A JSON file containing information about the wastewater or fecal sludge treatment plant. This file must be in a very specific format and can be created using the K2P Sketcher Tool (http://tools.waterpathogens.org/sketcher/)}
}
\description{
This function predicts the pathogen log reduction value for a wastewater or fecal sludge treatment plant sketched using the K2P Sketcher Tool (http://tools.waterpathogens.org/sketcher/)
}
\examples{
getLRV(mySketch="http://data.waterpathogens.org/dataset/afafd87c-b592-44c5-bb42-0a5a8415e54b/resource/83057ee9-402d-4b9b-8ab3-92053ee94c63/download/lubigisewageandfecalsludgetreatmentsystemv4.json",pathogenType="Virus",inFecalSludge=10000000000,inSewage=10000000000)

getLRV(mySketch="http://data.waterpathogens.org/dataset/bd53bbc6-a8f3-4d10-95b6-35f36e274b3c/resource/5cf1c3e0-a28c-4293-81e9-12586dd8941c/download/kirinyawastewatertreatmentplant-jinja6.json",pathogenType="Virus",inFecalSludge=10000000000,inSewage=10000000000)

}
\keyword{pathogens}
