% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPathogenFlows.R
\name{plotPathogenFlows}
\alias{plotPathogenFlows}
\title{The plotPathogenFlows function}
\usage{
plotPathogenFlows(sketcherResults)
}
\arguments{
\item{sketcherResults}{A data frame output from the getLRV function}
}
\description{
This function plots a Sankey flow chart showing pathogen loadings through a system sketched using the K2P Sketcher Tool (http://tools.waterpathogens.org/sketcher/)
}
\examples{
plotPathogenFlows(getLRV(mySketch="http://data.waterpathogens.org/dataset/afafd87c-b592-44c5-bb42-0a5a8415e54b/resource/83057ee9-402d-4b9b-8ab3-92053ee94c63/download/lubigisewageandfecalsludgetreatmentsystemv4.json",pathogenType="Virus",inFecalSludge=10000000000,inSewage=10000000000))

}
\keyword{pathogens}
