% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformMihalisData.R
\name{t_JMP}
\alias{t_JMP}
\title{The t_JMP function}
\usage{
t_JMP(contx)
}
\arguments{
\item{contx}{Either "urban" or "rural" context}
}
\description{
This function transforms country-level sanitation data from the UNICEF/WHO Joint Monitoring Program and produces it in a format that can be used directly by the getLoadings function.
}
\examples{
t_JMP("urban")


}
\keyword{pathogens}
