% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformJMPnat.R
\name{t_JMPnat}
\alias{t_JMPnat}
\title{The t_JMP function}
\usage{
t_JMPnat(context = "national", myRegions = "all")
}
\arguments{
\item{context}{Either "urban" or "rural" context}
}
\description{
This function transforms national or subnational sanitation data from the UNICEF/WHO Joint Monitoring Program and produces it in a format that can be used directly by the getLoadings function.
}
\examples{
t_JMP("urban")


}
\keyword{pathogens}
