% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformJMPsubnat.R
\name{t_JMPsubnat}
\alias{t_JMPsubnat}
\title{The t_JMP function}
\usage{
t_JMPsubnat(contx, myRegions = "all")
}
\arguments{
\item{contx}{Either "urban" or "rural" context}
}
\description{
This function transforms national or subnational sanitation data from the UNICEF/WHO Joint Monitoring Program and produces it in a format that can be used directly by the getLoadings function.
}
\examples{
t_JMP("urban")


}
\keyword{pathogens}
