% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoadings.R
\name{getLoadings}
\alias{getLoadings}
\title{The getLoadings function}
\usage{
getLoadings(
  inputDF = read.csv("data/input_file_new_kla_div_20200727.csv"),
  pathogenType = "Virus"
)
}
\arguments{
\item{inputDF}{A dataframe containing your onsite sanitation data. An example template can be found at http://data.waterpathogens.org/}

\item{pathogenType}{Specify either "Virus","Bacteria","Protozoa", or "Helminth"}
}
\description{
This function predicts the pathogen loadings from onsite sanitation systems for data available through the UNICEF/WHO Joint Monitoring Program and provides an output that can be used directly by the Pathogen Mapping Tool.
}
\examples{
myOutput<-getLoadings(inputDF,pathogenType="Virus")
myOutput$output

$output
     gid excreted   to_groundwater   to_surface       retained_in_soil      decayed         In_Fecal_Sludge    In_Sewage        stillViable        Onsite_LRV     Onsite_PR
1    HND 3.63e+18   5.840389e+16     1.963546e+18     5.738629e+17          5.042712e+17    1.179983e+14       5.297977e+17     2.551866e+18       0.15           0.2970
2    UGA 1.67e+19   1.565005e+17     1.731193e+18     1.408505e+18          1.340068e+19    0.000000e+00       3.127723e+15     1.890822e+18       0.95           0.8868
}
\keyword{pathogens}
