% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRV.R
\name{getLRV}
\alias{getLRV}
\title{The getLRV function}
\usage{
getLRV(
 
    mySketch = "http://data.waterpathogens.org/dataset/a1423a05-7680-4d1c-8d67-082fbeb00a50/resource/e7852e8f-9603-4b19-a5fa-9cb3cdc63bb8/download/sketch_lubigi.json",
  pathogenType = "Virus",
  inFecalSludge = 1e+10,
  inSewage = 1e+10
)
}
\arguments{
\item{pathogenType}{Pathogen group of interest (Virus, Bacteria, Protozoa, Helminths)}

\item{inFecalSludge}{Number of pathogens conveyed each year to the treatment plant in fecal sludge}

\item{inSewage}{Number of pathogens conveyed each year to the treatment plant in sewerage}

\item{sketch}{A JSON file containing information about the wastewater or fecal sludge treatment plant. This file must be in a very specific format and can be created using the K2P Sketcher Tool (http://tools.waterpathogens.org/sketcher/)}
}
\description{
This function predicts the pathogen log reduction value for a wastewater or fecal sludge treatment plant sketched using the K2P Sketcher Tool (http://tools.waterpathogens.org/sketcher/)
}
\examples{
getLRV(mySketch="http://data.waterpathogens.org/dataset/a1423a05-7680-4d1c-8d67-082fbeb00a50/resource/e7852e8f-9603-4b19-a5fa-9cb3cdc63bb8/download/sketch_lubigi.json",pathogenType="Virus",inFecalSludge=10000000000,inSewage=10000000000)



}
\keyword{pathogens}
