% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRV.R
\name{getLRV}
\alias{getLRV}
\title{The getLRV function}
\usage{
getLRV(
  mySketch = "data/lubigisewageandfecalsludgetreatmentsystem2.json",
 
    myLRVdata = "http://data.waterpathogens.org/dataset/eda3c64c-479e-4177-869c-93b3dc247a10/resource/9e172f8f-d8b5-4657-92a4-38da60786327/download/treatmentdata.csv",
  pathogenType = "Virus",
  inFecalSludge = 1e+10,
  inSewage = 1e+10
)
}
\arguments{
\item{mySketch}{A JSON file containing information about the wastewater or fecal sludge treatment plant. This file must be in a very specific format and can be created using the K2P Sketcher Tool (http://tools.waterpathogens.org/sketcher/)}

\item{myLRVdata}{A CSV file specifying data to use for fitting regression models that are used to predict pathogen reduction values in your sketched system}

\item{pathogenType}{Pathogen group of interest (Virus, Bacteria, Protozoa, Helminth)}

\item{inFecalSludge}{Number of pathogens conveyed each year to the treatment plant in fecal sludge}

\item{inSewage}{Number of pathogens conveyed each year to the treatment plant in sewerage}
}
\description{
This function predicts the pathogen log reduction value for a wastewater or fecal sludge treatment plant sketched using the K2P Sketcher Tool (http://tools.waterpathogens.org/sketcher/)
}
\examples{
getLRV(mySketch="data/lubigisewageandfecalsludgetreatmentsystem2.json",pathogenType="Virus",myLRVdata="http://data.waterpathogens.org/dataset/eda3c64c-479e-4177-869c-93b3dc247a10/resource/9e172f8f-d8b5-4657-92a4-38da60786327/download/treatmentdata.csv",pathogenType="Virus",inFecalSludge=10000000000,inSewage=10000000000)


}
\keyword{pathogens}
